Changes to DPDK code to print decimal counters; changed sketch counter sizes from 8 bit to 32 bit; enabled promiscuous mode

+++ sketch/CM.h
@@ -3,7 +3,7 @@
 
 #include "sketch.h"
 
-typedef uint8_t Value;
+typedef uint32_t Value;
 
 template<typename Key>
 struct CM_Entry{
